+++ ogre2/include/ignition/rendering/ogre2/Ogre2Visual.hh
@@ -41,6 +41,9 @@ namespace ignition
       public: virtual void SetVisible(bool _visible) override;
 
       // Documentation inherited.
+      public: virtual void SetVisibilityFlags(uint32_t _flags) override;
+
+      // Documentation inherited.
       protected: virtual GeometryStorePtr Geometries() const override;
 
       // Documentation inherited.